Use esbuild to bundle up the code before deploying to AWS Lambda
esbuild only deploys the code and dependencies needed to run the Lambda
function, which reduces the deployment size and allows all of the code
(including the AWS CDK script) to be placed within a single npm module.

See https://docs.aws.amazon.com/cdk/api/latest/docs/aws-lambda-nodejs-readme.html#reference-project-architecture
