+++ .gitignore
@@ -1,2 +1,6 @@
 node_modules
 .vscode
+
+# CDK asset staging directory
+.cdk.staging
+cdk.out
