--- infra/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "awscostbot-infra",
-  "version": "0.1.0",
-  "bin": {
-    "infra": "src/infra.js"
-  },
-  "scripts": {
-    "cdk": "cdk",
-    "bootstrap": "cdk bootstrap"
-  },
-  "dependencies": {
-    "@aws-cdk/aws-events": "1.106.1",
-    "@aws-cdk/aws-events-targets": "1.106.1",
-    "@aws-cdk/aws-iam": "1.106.1",
-    "@aws-cdk/aws-lambda": "1.106.1",
-    "@aws-cdk/aws-logs": "1.106.1",
-    "@aws-cdk/core": "1.106.1"
-  },
-  "devDependencies": {
-    "aws-cdk": "1.106.1"
-  }
-}
