--- lambda/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "awscostbot-lambda",
-  "version": "2.0.0",
-  "description": "Lambda function for aws-cost-bot",
-  "main": "src/index.js",
-  "license": "Apache-2.0",
-  "scripts": {
-    "start": "node src/index.js"
-  },
-  "dependencies": {
-    "@slack/web-api": "5.15.0",
-    "aws-cost-explorer": "0.1.3",
-    "aws-sdk": "2.814.0",
-    "axios": "0.21.2",
-    "js-currency-converter": "2.0.0",
-    "request": "2.88.2",
-    "discord.js": "12.5.3"
-  }
-}
