+++ lib/index.lambda.js
@@ -273,15 +273,3 @@ const generateMessage = (accountCosts, todaysConversionRate) => {
   });
   return message;
 };
-
-if (require.main === module) {
-  process.env.DRY_RUN = process.env.DRY_RUN || "true";
-
-  const event = {};
-  const context = {
-    done: (...args) => {
-      console.log(`context.done(${args.map(arg => String(arg)).join(", ")})`);
-    }
-  };
-  exports.handler(event, context).catch(e => console.error(e));
-}
