+++ package.json
@@ -1,16 +1,42 @@
 {
   "name": "awscostbot",
   "version": "1.0.0",
-  "description": "Provides visiblity on AWS costs",
+  "description": "Provides visibility on AWS costs",
   "main": "index.js",
   "scripts": {
-    "start": "npm start --prefix lambda",
-    "postinstall": "npm install --prefix lambda && npm install --prefix infra",
-    "deploy": "npm ci --prod --prefix lambda && npm ci --prefix infra && npm run --prefix infra bootstrap && npm run --prefix infra cdk -- deploy"
+    "start": "node bin/start.js",
+    "cdk": "cdk",
+    "bootstrap": "cdk bootstrap",
+    "deploy": "npm run lint && cdk deploy",
+    "lint": "tsc --noEmit"
   },
   "keywords": [
     "aws"
   ],
   "author": "neisha",
-  "license": "None"
+  "license": "None",
+  "dependencies": {
+    "@aws-cdk/aws-events": "1.132.0",
+    "@aws-cdk/aws-events-targets": "1.132.0",
+    "@aws-cdk/aws-iam": "1.132.0",
+    "@aws-cdk/aws-lambda": "1.132.0",
+    "@aws-cdk/aws-lambda-nodejs": "1.132.0",
+    "@aws-cdk/aws-logs": "1.132.0",
+    "@aws-cdk/core": "1.132.0",
+    "@slack/web-api": "5.15.0",
+    "aws-cost-explorer": "0.1.3",
+    "aws-sdk": "2.1032.0",
+    "axios": "0.24.0",
+    "discord.js": "12.5.3"
+  },
+  "devDependencies": {
+    "@types/node": "16.11.8",
+    "aws-cdk": "1.132.0",
+    "esbuild": "0.13.14",
+    "ts-node": "10.4.0",
+    "typescript": "4.5.2"
+  },
+  "engines": {
+    "node": ">=14"
+  }
 }
